% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cps_data_dir}
\alias{cps_data_dir}
\title{Get the cpsvote data directory}
\usage{
cps_data_dir()
}
\value{
A character string giving the expanded data directory path.
}
\description{
Returns the path where downloaded CPS data files are stored. The default is
\verb{~/cps_data}, which expands to your home directory on any platform.
}
\details{
To use a different location persistently, add this to your \code{.Rprofile}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{options(cpsvote.datadir = "~/my/path/to/cps_data")
}\if{html}{\out{</div>}}

Any path set this way will be used automatically by \code{\link[=cps_load_basic]{cps_load_basic()}},
\code{\link[=cps_read]{cps_read()}}, and \code{\link[=cps_download_data]{cps_download_data()}} without needing to pass \code{datadir}
each time.
}
\seealso{
\code{\link[=cps_load_basic]{cps_load_basic()}}, \code{\link[=cps_read]{cps_read()}}, \code{\link[=cps_download_data]{cps_download_data()}}
}
