% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cps_docs_dir}
\alias{cps_docs_dir}
\title{Get the cpsvote documentation directory}
\usage{
cps_docs_dir()
}
\value{
A character string giving the expanded documentation directory path.
}
\description{
Returns the path where downloaded CPS technical documentation files are
stored. The default is \verb{~/cps_docs}, which expands to your home directory
on any platform.
}
\details{
To use a different location persistently, add this to your \code{.Rprofile}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{options(cpsvote.docsdir = "~/my/path/to/cps_docs")
}\if{html}{\out{</div>}}

Any path set this way will be used automatically by \code{\link[=cps_download_docs]{cps_download_docs()}}
without needing to pass \code{path} each time.
}
\seealso{
\code{\link[=cps_download_docs]{cps_download_docs()}}
}
