% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_reweight_turnout.R
\name{cps_reweight_turnout}
\alias{cps_reweight_turnout}
\title{apply weight correction for voter turnout}
\usage{
cps_reweight_turnout(data)
}
\arguments{
\item{data}{the input data set, containing columns \code{YEAR}, \code{STATE}, and
\code{hurachen_turnout}}
}
\description{
This function applies the turnout correction recommended by Hur & Achen
(2013). The data set containing the scaling factor is \code{cpsvote::cps_reweight}.
}
\examples{
cps_reweight_turnout(cps_recode_vote(cps_refactor(cps_label(cps_2016_10k))))

}
