% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_cps.R
\name{label_cps}
\alias{label_cps}
\title{Apply factor levels to raw CPS data}
\usage{
label_cps(data, factors = cps_factors, names_col = "new_name")
}
\arguments{
\item{data}{The raw CPS data that factors should be applied to}

\item{factors}{A data frame containing the label codes to be applied}

\item{names_col}{Which column of \code{factors} contains the column names of \code{data}}
}
\value{
CPS data with factor labels in place of the raw numeric data
}
\description{
The CPS publishes their data in a numeric format, with a separate
PDF codebook (not machine readable) describing factor values. This function
labels the raw numeric CPS data according to a supplied factor key. Codes
that appear in a given yer and are not included in \code{factors} will be
supplanted with \code{NA}.
}
