% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_label.R
\name{na_ifin}
\alias{na_ifin}
\title{vectorized \code{na_if}}
\usage{
na_ifin(x, y)
}
\arguments{
\item{x}{the vector to be checked}

\item{y}{the values which should be replaced with NA}
}
\description{
vectorized \code{na_if}
}
