% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cps.R
\name{read_cps}
\alias{read_cps}
\title{Read in CPS data}
\usage{
read_cps(data_dir = "cps_data", years = seq(1994, 2018, 2),
  factored = TRUE, catalog = "default", join_dfs = factored,
  combine_factors = join_dfs)
}
\arguments{
\item{data_dir}{The folder where the CPS data files live. These files should
follow a naming scheme that contains the 4-digit year of the results in
question, and have a ".zip" or ".gz" extension.}

\item{years}{Which years to read in.}

\item{factored}{Whether the data (in numeric form) should be converted to
the equivalent factor values or not. This will also rename columns according
to the \code{catalog} argument.
#' @param catalog Which columns to read, and how to assign factor labels. The
default value is "default", which reads from the list \code{cpsvote:::fwf_key}.}

\item{join_dfs}{Whether to combine all of the years into a single data frame,
or leave them as a list of data frames. This will default to the value of
\code{factored}, because the raw column names represent different questions in
different years and should not be naively joined otherwise.}

\item{combine_factors}{Whether to merge the changing factor levels over time
into one consistent set of factors. This will default to the value of
\code{join_dfs}, and is only set up for the default data on the complete set of
years (even-numbered years from 1994-2018).}
}
\description{
Load data from the Current Population Survey
}
