% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cps.R
\name{read_year}
\alias{read_year}
\title{Load a single CPS file}
\usage{
read_year(file, year = as.numeric(stringr::str_extract(file,
  "\\\\d{4}")), factored = TRUE, catalog = "default")
}
\arguments{
\item{file}{Where the fwf or zip (or gz) file for this year's data lives}

\item{year}{Which year is being read; defaults to 4-digit year in file name
#' @param factored Whether the data (in numeric form) should be converted to
the equivalent factor values or not. This will also rename columns according
to the \code{catalog} argument.}

\item{catalog}{Which columns to read, and how to assign factor labels. The
default value is "default", which reads from the list \code{cpsvote:::fwf_key}.}
}
\description{
Read a single year of data from the CPS
}
