% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps_reweight.R
\name{cps_reweight}
\alias{cps_reweight}
\title{Attach proper weights to calculate voter turnout}
\usage{
cps_reweight(data)
}
\arguments{
\item{data}{The data set, containing a column of original weights}
}
\description{
Attach proper weights to calculate voter turnout
}
\details{
This function is designed to reweight the CPS VRS data for properly
calculated turnout values, according to Achen and Hur (2013). Without
reweighting, any estimate of voter turnout will be inaccurate. For more
information on this reweighting, see \code{vignette("cps")}.

Additionally, it is designed to join and weight according to the default
column names as used in the included \code{cps} data set - so it will throw an
error if the required column names are not present. For custom data sets,
load the provided data set \code{reweight} and join the two on year, state, and
reported voting. Then multiply the existing weight column by the reweighting
factor to obtain a correct weight to calculate turnout.
}
