% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cps.R
\name{read_cps}
\alias{read_cps}
\title{Read in CPS data}
\usage{
read_cps(
  dir = "cps_data",
  years = seq(1994, 2018, 2),
  cols = cps_cols,
  join_dfs = TRUE
)
}
\arguments{
\item{dir}{The folder where the CPS data files live. These files should
follow a naming scheme that contains the 4-digit year of the results in
question, and have a ".zip" or ".gz" extension.}

\item{years}{Which years to read in. Thie function will read data from files
in \code{dir} whose names contain these 4-digit years.
#' @param cols Which columns to read. This must be a data frame, with required
columns \code{start_pos},\code{end_pos}, and \code{year}. The default value is \code{cps_cols},
which reads from the list \code{cpsvote::cps_cols}. See \url{vignettes/read_specs.html}
for details about how to specify a different set of \code{cols}.}

\item{join_dfs}{Whether to combine all of the years into a single data frame,
or leave them as a list of data frames.}
}
\value{
a data frame, or list of data frames
}
\description{
Load multiple years of data from the Current Population Survey
}
